+++ src/rb_sys.rs
@@ -195,7 +195,7 @@ pub unsafe fn resume_error(e: Error) -> ! {
 }
 
 /// Convert [`magnus::Value`](Value) to [`rb_sys::VALUE`].
-#[deprecated(since = "0.3.3", note = "please use `Value::as_raw` instead")]
+#[deprecated(since = "0.4.0", note = "please use `Value::as_raw` instead")]
 pub fn raw_value(val: Value) -> VALUE {
     val.as_raw()
 }
@@ -207,13 +207,13 @@ pub fn raw_value(val: Value) -> VALUE {
 /// You must only supply a valid [`VALUE`] obtained from [rb-sys](rb_sys) to
 /// this function. Using a invalid [`Value`] produced from this function will
 /// void all saftey guarantees provided by Magnus.
-#[deprecated(since = "0.3.3", note = "please use `Value::from_raw` instead")]
+#[deprecated(since = "0.4.0", note = "please use `Value::from_raw` instead")]
 pub unsafe fn value_from_raw(val: VALUE) -> Value {
     Value::from_raw(val)
 }
 
 /// Convert [`magnus::value::Id`](Id) to [`rb_sys::ID`].
-#[deprecated(since = "0.3.3", note = "please use `Id::as_raw` instead")]
+#[deprecated(since = "0.4.0", note = "please use `Id::as_raw` instead")]
 pub fn raw_id(id: Id) -> ID {
     id.as_raw()
 }
@@ -225,7 +225,7 @@ pub fn raw_id(id: Id) -> ID {
 /// You must only supply a valid [`ID`] obtained from [rb-sys](rb_sys) to this
 /// function. Using a invalid [`Id`] produced from this function will void all
 /// saftey guarantees provided by Magnus.
-#[deprecated(since = "0.3.3", note = "please use `Id::from_raw` instead")]
+#[deprecated(since = "0.4.0", note = "please use `Id::from_raw` instead")]
 pub unsafe fn id_from_raw(id: ID) -> Id {
     Id::from_raw(id)
 }